query for hosts with trailing digits
This results in containers app1 and app2 having both their records listed for app.example.com.
At this point however it is only supported as long as the app instances are running on different remotes.
Locally exact matches are preferred and beyond that the first record is chosen which is LXD implementation dependent.

Signed-off-by: benaryorg <binary@benary.org>