pipe backend support
This change intends to support both the existing unix domain socket backend as well as the PowerDNS piped backend (actually the remote backend's pipe protocol, important difference).
While it does this it also adds some major refactoring, which also moves the queue creation to a different spot, again, which *should* in theory mean that queues are neither thrashed nor shared.
A breaking change will be required for this since it changes the command line usage (it does subcommands now).
The resolver/responder part is also refactored into its own process and must now be run separately.
This means that on every node of the "cluster" you have to run the responder subcommand and on all DNS server instances run either the unix subcommand or let PowerDNS call the pipe subcommand directly.

Signed-off-by: benaryorg <binary@benary.org>