http: max-connections paramater
This allows setting a maximum connection limit *per worker*.
With this you can limit the impact of the API being hit with a lot of requests, since every request will spawn a sudo process.
Using HTTPS with certificate based authentication would also solve this but requires significant additional changes and would only partially mitigate this (and offload the problem to the hypervisor directly), so at least vague rate-limiting can be used for now.

Signed-off-by: benaryorg <binary@benary.org>