reqwest: native certificates with rustls
First of all this makes it easier to test with self-signed certs, second it dispenses with having to update the bundled certs, and it's just generally eeasier to work with.
The project uses rustls statically because that way it's easier to package this application, especially when OpenSSL vs. LibreSSL is a thing, but bundling certs was never intentionally.
However this *is* a breaking change and thus will get its major version bumped.

Signed-off-by: benaryorg <binary@benary.org>